Rename `ForgotPinResult` -> `ResetPinResult`
