+++ app/src/main/java/org/simple/clinic/forgotpin/confirmpin/ForgotPinConfirmPinScreenController.kt
@@ -10,10 +10,10 @@ import org.simple.clinic.ReplayUntilScreenIsDestroyed
 import org.simple.clinic.ReportAnalyticsEvents
 import org.simple.clinic.facility.FacilityRepository
 import org.simple.clinic.patient.PatientRepository
-import org.simple.clinic.user.resetpin.ForgotPinResult.NetworkError
-import org.simple.clinic.user.resetpin.ForgotPinResult.Success
-import org.simple.clinic.user.resetpin.ForgotPinResult.UnexpectedError
-import org.simple.clinic.user.resetpin.ForgotPinResult.UserNotFound
+import org.simple.clinic.user.resetpin.ResetPinResult.NetworkError
+import org.simple.clinic.user.resetpin.ResetPinResult.Success
+import org.simple.clinic.user.resetpin.ResetPinResult.UnexpectedError
+import org.simple.clinic.user.resetpin.ResetPinResult.UserNotFound
 import org.simple.clinic.user.User.LoggedInStatus.*
 import org.simple.clinic.user.UserSession
 import org.simple.clinic.widgets.UiEvent
