--- app/src/main/java/org/simple/clinic/user/resetpin/ForgotPinResult.kt
@@ -1,8 +1,0 @@
-package org.simple.clinic.user.resetpin
-
-sealed class ForgotPinResult {
-  object Success : ForgotPinResult()
-  object NetworkError : ForgotPinResult()
-  object UserNotFound : ForgotPinResult()
-  data class UnexpectedError(val cause: Throwable? = null) : ForgotPinResult()
-}
